backend/requirements.txt
absl-py==2.4.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
astunparse==1.6.3
billiard==4.2.4
celery==5.6.3
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
fastapi==0.136.1
flatbuffers==25.12.19
fonttools==4.62.1
gast==0.7.0
google-pasta==0.2.0
greenlet==3.5.0
grpcio==1.80.0
gunicorn==26.0.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
jiter==0.14.0
joblib==1.5.3
keras==3.14.1
kiwisolver==1.5.0
kombu==5.6.2
libclang==18.1.1
markdown-it-py==4.2.0
matplotlib==3.10.9
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
numpy==2.4.4
openai==2.36.0
opt_einsum==3.4.0
optree==0.19.1
packaging==26.2
pandas==3.0.2
pillow==12.2.0
prompt_toolkit==3.0.52
protobuf==7.34.1
pydantic==2.13.4
pydantic_core==2.46.4
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
redis==7.4.0
requests==2.33.1
rich==15.0.0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
starlette==1.0.0
tensorflow==2.21.0
termcolor==3.3.0
threadpoolctl==3.6.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
tzlocal==5.3.1
urllib3==2.7.0
uvicorn==0.46.0
vine==5.1.0
wcwidth==0.7.0
wrapt==2.1.2
